backend/sgc/requirements.txt
asgiref==3.2.10
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.1
drf-spectacular==0.9.13
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==20.4
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
